+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 # Changelog
 
+## v0.15.0 (Unreleased)
+
+- Added `Exception::Custom` and `Exception::new`.
+
+### Breaking Changes
+
+- Removed `TryFrom<u8>` and `#[repr(u8)]` for Exception.
+
 ## v0.14.1 (2024-09-10)
 
 - Implement `Report Server ID` (function code 17).